feat: combined evaluation functions (nlp and llm based) into one wine multiple modes (dockerfile to be updated)
