+++ app/nlp_testing_nltk.py
@@ -1,5 +1,5 @@
-import nltk
-print(nltk.data.path)
-nltk.download()
+# import nltk
+# print(nltk.data.path)
+# nltk.download()
 # If zip packages cannot be unzipped or error from the downloader above, then download packages from online https://www.nltk.org/nltk_data/
 # Need to check with the command above where the zip packages should go in folder /Users/<username>/nltk_data/...
